+++ Cargo.toml
@@ -26,10 +26,9 @@ mysql = []
 redis_compression = ["dep:zstd"]
 
 [dependencies]
-gsm = { path = "crates/gsm" }
+gsm = { path = "crates/gsm", features = ["loader", "s3"] }
 aws-config = { version = "1.5.5" }
 aws-sdk-kms = { version = "1.40.0", optional = true }
-aws-sdk-s3 = { version = "1" }
 aws-sdk-sesv2 = { version = "1" }
 lettre = { version = "0.11", features = ["tokio1", "tokio1-native-tls"] }
 base64 = "0.22.1"
