+++ crates/gsm/Cargo.toml
@@ -6,7 +6,19 @@ rust-version = "1.85.0"
 description = "Gateway Status Map — connector error codes mapped to routing decisions"
 license = "Apache-2.0"
 
+[features]
+default = []
+# Adds the static store, init(), options(), and lookup() helpers.
+loader = ["dep:once_cell", "dep:tracing"]
+# Adds S3 as a load source; implies loader.
+s3 = ["loader", "dep:aws-config", "dep:aws-sdk-s3"]
+
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
 csv = "~1.3"
 thiserror = "2.0"
+
+once_cell  = { version = "1.19",  optional = true }
+tracing    = { version = "0.1",   optional = true }
+aws-config = { version = "1.5.5", optional = true }
+aws-sdk-s3 = { version = "1",     optional = true }
